refactor: complete revamp of CLI commands
