+++ README.md
@@ -41,7 +41,7 @@ The CLI allow to load the JSON into a Webpage with the **open** command. The pag
 
 ## Requirements
 
-- [Node.js](https://nodejs.org/en/) version 16.0.0 or higher
+- [Node.js](https://nodejs.org/en/) version 16.8.0 or higher
 
 ## Getting Started
