+++ package.json
@@ -1,14 +1,15 @@
 {
   "name": "nsecure",
-  "version": "0.8.0",
+  "version": "0.9.0",
   "description": "Node.js security CLI",
   "main": "./index.js",
   "bin": {
     "node-secure": "./bin/index.js",
     "nsecure": "./bin/index.js"
   },
+  "type": "module",
   "engines": {
-    "node": ">=12.12.0"
+    "node": ">=16.1.0"
   },
   "scripts": {
     "eslint": "eslint bin src test",
@@ -88,19 +89,16 @@
     "@nodesecure/flags": "^1.0.0",
     "@nodesecure/i18n": "^1.1.1",
     "@nodesecure/npm-registry-sdk": "^1.2.0",
-    "@nodesecure/scanner": "^1.0.0",
+    "@nodesecure/scanner": "^1.1.0",
     "@nodesecure/vuln": "^1.2.0",
     "@polka/send-type": "^0.5.2",
     "@slimio/async-cli-spinner": "^0.5.2",
-    "cacache": "^15.3.0",
     "cliui": "^7.0.4",
     "dotenv": "^10.0.0",
     "filenamify": "^5.0.1",
     "kleur": "^4.1.4",
-    "make-promises-safe": "^5.1.0",
     "ms": "^2.1.3",
     "open": "^8.2.1",
-    "pacote": "^11.3.5",
     "polka": "^0.5.2",
     "qoa": "^0.2.0",
     "sade": "^1.7.4",
