+++ src/commands/vulnerability.js
@@ -1,0 +1,29 @@
+// Import Third-party Dependencies
+import * as vuln from "@nodesecure/vuln";
+import * as i18n from "@nodesecure/i18n";
+
+import ms from "ms";
+import Spinner from "@slimio/async-cli-spinner";
+import kleur from "kleur";
+
+export async function hydrate() {
+  const strategy = await vuln.setStrategy(vuln.strategies.SECURITY_WG, {
+    hydrateDatabase: false
+  });
+
+  await strategy.deleteDatabase();
+
+  const spinner = new Spinner({
+    text: kleur.white().bold(i18n.getToken("cli.commands.hydrate_db.running", kleur.yellow().bold("nodejs security-wg")))
+  }).start();
+
+  try {
+    await strategy.hydrateDatabase();
+
+    const elapsedTime = cyan(ms(Number(spinner.elapsedTime.toFixed(2))));
+    spinner.succeed(white().bold(i18n.getToken("cli.commands.hydrate_db.success", elapsedTime)));
+  }
+  catch (err) {
+    spinner.failed(err.message);
+  }
+}
