+++ src/utils.js
@@ -9,3 +9,10 @@ export function formatBytes(bytes, decimals) {
 
   return parseFloat((bytes / Math.pow(1024, id)).toFixed(dm)) + " " + sizes[id];
 }
+
+export function locationToString(location) {
+  const start = `${location[0][0]}:${location[0][1]}`;
+  const end = `${location[1][0]}:${location[1][1]}`;
+
+  return `[${start}] - [${end}]`;
+}
